# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:4444",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "lms-backend",
    "description": "",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "paths": {
...
    "/v1/auth/kakao/login": {
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KakaoLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcludePasswordUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcludePasswordUserDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/signup": {
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcludePasswordUserDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "409": {
            "description": "user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse409"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/role": {
      "patch": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExcludePasswordUserDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "user not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/category": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "withChildren",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "required": true
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "required": false
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "name": "orderBy",
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "const": "asc"
                },
                {
                  "const": "desc"
                }
              ]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryWithChildrenDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      }
    },
    "/v1/category/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/CategoryDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "403": {
            "description": "category has courses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse403"
                }
              }
            }
          },
          "404": {
            "description": "category not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ExcludePasswordUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "emailVerified": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "role": {
            "oneOf": [
              {
                "const": "user"
              },
              {
                "const": "teacher"
              },
              {
                "const": "manager"
              },
              {
                "const": "admin"
              }
            ]
          },
          "image": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "date-time"
              }
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "email",
          "emailVerified",
          "role",
          "image",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "KakaoLoginDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "gender": {
            "oneOf": [
              {
                "const": "male"
              },
              {
                "const": "female"
              }
            ]
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "birthdayType": {
            "oneOf": [
              {
                "const": "SOLAR"
              },
              {
                "const": "LUNAR"
              }
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "connectingInformation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "duplicationInformation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "name",
          "email",
          "gender",
          "birthDate",
          "birthdayType",
          "phoneNumber",
          "connectingInformation",
          "duplicationInformation"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TypeGuardErrorany": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "expected": {
            "type": "string"
          },
          "value": {},
          "fake_expected_typed_value_": {},
          "name": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string"
          }
        },
        "required": [
          "method",
          "expected",
          "value",
          "name",
          "message"
        ]
      },
      "IErrorResponse404": {
        "type": "object",
        "properties": {
          "statusCode": {
            "const": 404
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "SignUpUserDto": {
        "type": "object",
        "properties": {
          "userCreateParams": {
            "$ref": "#/components/schemas/CreateUserDto"
          },
          "infoCreateParams": {
            "$ref": "#/components/schemas/CreateUserInfoDto"
          },
          "accountCreateParams": {
            "$ref": "#/components/schemas/CreateUserAccountDto"
          }
        },
        "required": [
          "userCreateParams",
          "infoCreateParams",
          "accountCreateParams"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "role": {
            "oneOf": [
              {
                "const": "user"
              },
              {
                "const": "teacher"
              },
              {
                "const": "manager"
              },
              {
                "const": "admin"
              }
            ]
          }
        },
        "required": [
          "displayName",
          "email",
          "password",
          "role"
        ]
      },
      "CreateUserInfoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "birthDate": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "date"
              }
            ]
          },
          "gender": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "phoneNumber": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "connectingInformation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "duplicationInformation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "name",
          "birthDate",
          "gender",
          "phoneNumber",
          "connectingInformation",
          "duplicationInformation"
        ]
      },
      "CreateUserAccountDto": {
        "type": "object",
        "properties": {
          "providerType": {
            "oneOf": [
              {
                "const": "email"
              },
              {
                "const": "kakao"
              }
            ]
          },
          "providerId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "providerType",
          "providerId"
        ]
      },
      "IErrorResponse409": {
        "type": "object",
        "properties": {
          "statusCode": {
            "const": 409
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "oneOf": [
              {
                "const": "user"
              },
              {
                "const": "teacher"
              },
              {
                "const": "manager"
              },
              {
                "const": "admin"
              }
            ]
          }
        },
        "required": [
          "id",
          "role"
        ]
      },
      "CategoryWithChildrenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uuid"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryWithChildrenDto"
            }
          }
        },
        "required": [
          "id",
          "parentId",
          "name",
          "description",
          "children"
        ]
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uuid"
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "id",
          "parentId",
          "name",
          "description"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "parentId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uuid"
              }
            ]
          }
        },
        "required": [
          "name",
          "description",
          "parentId"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "parentId": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uuid"
              }
            ]
          }
        }
      },
      "IErrorResponse403": {
        "type": "object",
        "properties": {
          "statusCode": {
            "const": 403
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}